moved reconciler files to discovery
