+++ internal/controller/discovery/message_processor.go
@@ -1,4 +1,4 @@
-package reconciler
+package discovery
 
 import (
 	"context"
